requirements.txt
annotated-types==0.7.0
anyio==4.12.1
APScheduler==3.11.2
certifi==2026.1.4
distro==1.9.0
exceptiongroup==1.3.1
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
jiter==0.13.0
openai==2.20.0
pydantic==2.12.5
pydantic_core==2.41.5
python-telegram-bot==22.5
sniffio==1.3.1
tqdm==4.67.3
typing-inspection==0.4.2
typing_extensions==4.15.0
tzlocal==5.3.1
python-telegram-bot==22.5

